+++ package.json
@@ -47,6 +47,7 @@
     "react-dev-utils": "^6.1.1",
     "react-dom": "^16.6.3",
     "react-icons": "^3.2.2",
+    "react-router-dom": "^4.3.1",
     "react-svg-piechart": "^2.1.1",
     "reactstrap": "^6.5.0",
     "resolve": "1.8.1",
