--- src/components/App/App.scss
@@ -1,20 +1,0 @@
-.app-container {
-  margin: 50px 0;
-  .loading {
-    text-align: center;
-  }
-  .participant-container {
-    PieChart {
-      margin-top: 25px;
-    }
-    .participant-list-item {
-      display: flex;
-      align-items: center;
-      span:first-child {
-        flex: 1;
-      }
-    }
-  }
-  .thread-container {
-  }  
-}
